requirements.txt
pynput
paho-mqtt